fix (SpecialPackageSection): change getRatingComponentList helper to RatingComponentList component
